+++ build.clj
@@ -2,8 +2,8 @@
   (:refer-clojure :exclude [test])
   (:require [clojure.tools.build.api :as b]))
 
-(def lib 'io.github.com/konpy2)
-(def version "0.2.9")
+(def lib 'io.github.hkimjp/konpy2)
+(def version "0.2.13")
 (def main 'hkimjp.konpy2.main)
 (def class-dir "target/classes")