+++ src/hkimjp/konpy2/comments.clj
@@ -1,39 +1,42 @@
 (ns hkimjp.konpy2.comments
   (:require
-   [ring.util.response :as resp]
    [taoensso.telemere :as t]
    [hkimjp.datascript :as ds]
-   [hkimjp.konpy2.response :refer [page hx]]
+   [hkimjp.konpy2.response :refer [page hx redirect]]
    [hkimjp.konpy2.util :refer [now]]))
 
-(defn comment
-  "return comment eid=e"
-  [{{:keys [e]} :path-params}]
-  (let [e (parse-long e)
-        comm (-> (ds/pl e)
-                 :comment)]
-    (t/log! {:level :info :id "comment" :data comm})
-    (hx [:div comm])))
+(def ^:private comments-to
+  '[:find ?e ?author
+    :in $ ?to
+    :where
+    [?e :comment/status "yes"]
+    [?e :author ?author]
+    [?e :to ?to]])
 
-(def ^:private problem-eid '[:find ?pid
-                             :in $ ?cid
-                             :where
-                             [?cid :to ?pid]])
+; (defn div-comments
+;   "returns comments sent to `e`"
+;   [e]
+;   (t/log! {:level :info :id "comments-div" :data e})
+;   [:div
+;    (for [[e author] (ds/qq comments-to e)]
+;      [:button.pr-4 author])])
 
-;; (-> (ds/qq problem-eid 23) first)
-
+; pid をもらってこないと戻るページがない
 (defn post-comment
   "send comments to `e`.
    returns clickable commenter's list"
-  [{{:keys [author e comment]} :params}]
-  (t/log! {:level :info :id "post-comment" :data (parse-long e)})
-  (let [e (parse-long e)]
-    (try
-      (ds/put! {:comment/status "yes"
-                :author author
-                :to e
-                :comment comment
-                :updated (now)})
-      (resp/redirect (str "/k/problem/" (-> (ds/qq problem-eid e) ffirst)))
-      (catch Exception ex
-        (t/log! :error ex)))))
+  [{{:keys [to author comment pid]} :params}]
+  (t/log! {:level :info
+           :id    "post-comment"
+           :data  {:to to :author author :comment comment :pid pid}})
+  (ds/put! {:comment/status "yes"
+            :author author
+            :to (parse-long to)
+            :comment comment
+            :updated (now)})
+  (redirect (str "/k/problem/" pid)))
+
+(defn show-comment [{{:keys [e]} :path-params}]
+  (hx [:div (:comment  (ds/pl (parse-long e)))]))
+
+; (show-comment {:path-params {:e "47"}})